refactor: move home page component
