+++ plugins/announcements/src/components/homeComponents/AnnouncementsCard/index.ts
